pants-plugins/oci/pants_backend_oci/requirements.txt
python-debian==0.1.49
